+++ Sources/Action/Action.swift
@@ -26,9 +26,6 @@ When this excuted via execute() or inputs subject, it passes its parameter to th
 public final class Action<Input, Element> {
     public typealias WorkFactory = (Input) -> Observable<Element>
 
-    public let _enabledIf: Observable<Bool>
-    public let workFactory: WorkFactory
-
     /// Inputs that triggers execution of action.
     /// This subject also includes inputs as aguments of execute().
     /// All inputs are always appear in this subject even if the action is not enabled.
@@ -48,7 +45,7 @@ public final class Action<Input, Element> {
     public let elements: Observable<Element>
 
     /// Whether or not we're currently executing. 
-    public let executing: Observable<Bool>
+    public let isExecuting: Observable<Bool>
 
     /// Observables returned by the workFactory.
     /// Useful for sending results back from work being completed
@@ -58,7 +55,7 @@ public final class Action<Input, Element> {
     /// Whether or not we're enabled. Note that this is a *computed* sequence
     /// property based on enabledIf initializer and if we're currently executing.
     /// Always observed on MainScheduler.
-    public let enabled: Observable<Bool>
+    public let isEnabled: Observable<Bool>
 
     private let disposeBag = DisposeBag()
 
@@ -75,11 +72,8 @@ public final class Action<Input, Element> {
         enabledIf: Observable<Bool> = Observable.just(true),
         workFactory: @escaping WorkFactory) {
 
-        self._enabledIf = enabledIf
-        self.workFactory = workFactory
-
         let enabledSubject = BehaviorSubject<Bool>(value: false)
-        enabled = enabledSubject.asObservable()
+        isEnabled = enabledSubject.asObservable()
 
         let errorsSubject = PublishSubject<Error>()
         errors = errorsSubject.asObservable()
@@ -88,7 +82,7 @@ public final class Action<Input, Element> {
         disabledErrors = disabledErrorSubject.asObservable()
 
         executionObservables = inputs
-            .withLatestFrom(enabled) { input, enabled in (input, enabled) }
+            .withLatestFrom(isEnabled) { input, enabled in (input, enabled) }
             .flatMap { input, enabled -> Observable<Observable<Element>> in
                 if enabled {
                     return Observable.of(workFactory(input)
@@ -104,7 +98,7 @@ public final class Action<Input, Element> {
         elements = executionObservables
             .flatMap { $0.catchError { _ in Observable.empty() } }
 
-        executing = executionObservables.flatMap {
+        isExecuting = executionObservables.flatMap {
                 execution -> Observable<Bool> in
                 let execution = execution
                     .flatMap { _ in Observable<Bool>.empty() }
@@ -118,7 +112,7 @@ public final class Action<Input, Element> {
             .share(replay: 1, scope: .forever)
 
         Observable
-            .combineLatest(executing, enabledIf) { !$0 && $1 }
+            .combineLatest(isExecuting, enabledIf) { !$0 && $1 }
             .bind(to: enabledSubject)
             .disposed(by: disposeBag)
     }
@@ -142,5 +136,18 @@ public final class Action<Input, Element> {
 			.disposed(by: disposeBag)
 
 		return subject.asObservable()
+    }
+}
+
+// MARK: Deprecated
+extension Action {
+    @available(*, deprecated, renamed: "isExecuting")
+    public var executing: Observable<Bool> {
+        return isExecuting
+    }
+
+    @available(*, deprecated, renamed: "isEnabled")
+    public var enabled: Observable<Bool> {
+        return isEnabled
     }
 }