+++ Tests/ActionTests/ActionTests.swift
@@ -82,11 +82,11 @@ class ActionTests: QuickSpec {
                     .bind(to: disabledErrors)
                     .disposed(by: disposeBag)
 
-				action.enabled
+				action.isEnabled
 					.bind(to: enabled)
 					.disposed(by: disposeBag)
 				
-				action.executing
+                action.isExecuting
 					.bind(to: executing)
 					.disposed(by: disposeBag)
 				
@@ -98,8 +98,8 @@ class ActionTests: QuickSpec {
 				action.inputs.subscribe().disposed(by: disposeBag)
 				action.elements.subscribe().disposed(by: disposeBag)
 				action.errors.subscribe().disposed(by: disposeBag)
-				action.enabled.subscribe().disposed(by: disposeBag)
-				action.executing.subscribe().disposed(by: disposeBag)
+				action.isEnabled.subscribe().disposed(by: disposeBag)
+                action.isExecuting.subscribe().disposed(by: disposeBag)
 				action.executionObservables.subscribe().disposed(by: disposeBag)
 			}